VaultKeeps render in vault view. Delete Auth set up.

+++ Repositories/VaultsRepository.cs
@@ -65,9 +65,9 @@ namespace keepr.Repositories
 
     }
 
-    public bool Delete(int id)
+    public bool Delete(int id, string userId)
     {
-      int success = _db.Execute("DELETE FROM vaults WHERE id = @id", new { id });
+      int success = _db.Execute("DELETE FROM vaults WHERE id = @id AND userId = @userId", new { id, userId });
       return success > 0;
     }
